Removing pragmas for memory alignment in HashTableEntry
The pragma pack is causing errors in certain environments.
The issue was that the memory block containing the HashTableEntries
was aligned by 1-byte, but the delete[] operations was relying on
4-byte alignment.

I was not able to see which toolchain/compiler option or other
pragma was causing this error. Therefore I suggest to remove this
option as soon as we are sure this option works on each platform.

Change-Id: Ic7421fc47f23abf9dd5bd447734cb4d24c298a6b