Issue 32: Adding support for iPhoneOS and iPhoneSimulatorOS
This patch adds support for both iPhoneOS and iPhoneSimulatorOS.

For the iPhoneOS running on iOS device like the iPhone the following
platforms are currently supported: ARMV7, ARMV7S, ARM64

Furthermore the iPhone Simulator gets supported. Here the target
platform x86 32 and x86 64 bit are available.

Notice: It is not yet possible to build the GoogleTest framework for the
iPhoneOS platforms.

Change-Id: I6a17fb28c623f2ba35f82dd0eb60e9872f24477f