Adding additional test for timeout on send
Till now in TcpSocketInputStreamTest there was a test which checked the
timeout on send in case the receiver is too slow to read all the data. The
consequence is that the TCP receive window gets 0 in that case and the
sender gets a timeout.
This test gets now moved to TcpSocketTest as it directly tests the send.

Change-Id: I175a711e35e266362aa6abbd080919a838d0893f