Improve used types in hash-related functionality
Previously the hash function always returned the native integer type.
So on 32 bit systems one would get a 32 bit hash e.g. for a string,
and a 64 bit hash on a 64 bit system. This gives good behaviour for
instance for hashmap (uses all nativly available bits). This default
behaviour is still valid, but now one can also explicitly specify
the integer return type and the number bits to be used within.
This can be used to guarantee the same hashvalue for a certain data
regardless of architecture integer size.

Change-Id: I424980870cef6df59fbe4687a8769f5fd0a22a0e