Added map container class.
Implementation of a map that stores elements of type T sorted by their keys of type Key.
NOTE: Not STL compatible.
Each element must have a key, that is unique in the map. For key comparison the "<" operator must be defined for the type Key.
Elements can be inserted, searched and removed from the map in O(log(n)).
The iterator can be used to navigate forward and backward to the next, previous element in the sorting order (O(log(n)).

Change-Id: Id055ef0989bcc11781c8aaa76cec1975adf585f8