valgrind: Correct cleanup of DltAdapter
DltAdapter is a singleton. The pointer for the singleton instance is
filled when DltAdapter is used the first time. It is never cleaned up
though. When using an actual instance instead of a pointer the compiler
cleans up. Trough lazy initialisation race conditions with parallel
access to constructor and getters can be avoided.

Change-Id: I1b9ca1fd781a3246c2425b3111d136af35e2bc1c