make mutex in ConsoleLogAppender static
If this variable is not static it exists seperately in every thread.
It can therefore happen that logMessage is called concurrently from
different threads. This previously led to segfaults under certain
circumstances. This way with the static variable there is only
one mutex for the process and all threads are sharing the mutex.
This way the logMessage cannot be entered twice from different threads.

Change-Id: I4259748a12fda0842920c260deef9e4c51c6efe3