Implement a timer utility is that can be used for the cyclic execution of a delegate. Create tests for the timer utility
All the created timers has to be initialized with a TimerManager, which
is responsible for executing the Timer.
When a timer is started it registers itself to the TimerManager, when
it is stopped, or no executions are left, it is unregistered.
The TimerManager will make sure that Timer::execute() is called
periodically.
Every TimerManager uses a different thread to support its timers.
The thread will be stopped after all the timers are unregistered.
The timers alone could not handle the thread: if the last timer was
destroyed inside of an executed delegate, then the currently executed
thread would be destroyed, so the execution would stop unexpectedly.

Change-Id: I5d363c5c4578496d50fdf88eb6c7fe417e2950cd