Adding specialization for hash calculation of shared_ptr
With the specialization for hash calculation, the internal pointer
of the shared_ptr is passed to the hash function.

Wihtout this specialization the "whole" shared_ptr is hashed.
Meaning the internal pointer as well the pointer to the metadata.
This is perferctly fine when you use the shared_ptr in a regular
way.
With the possibility to add your own deleter functor you are
able to create different shared pointer instances that
contain different metadata. As demonstrated in
HashSeveralSharedPtrWithUserDeleter

Change-Id: I44d1119e269601715f9e09ff49ab19c753f0a4fc