Remove SO_REUSEADDR if bind-ing to port 0
Port 0 means choose any free port and implies that the port is
distinct from any other port. That conflict with the possibility that
the OS might choose an existing port from the process and resuse it.

Fix by moving the socket option setter to bind where the port is
accessible and guard by check that a specific port is requested.

Change-Id: I3d2323b172c9c09e515e4d5d487814c740d7b40b